pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "orchid-python-api"
version = "2020.4.595"
description = "Defines and implements the Python API for Orchid*. (*Orchid is a mark of Reveal Energy Services, Inc.)"
authors = [ "Reveal Energy Services, Inc. <support@reveal-energy.com>",]
maintainers = [ "Reveal Energy Services, Inc. <support@reveal-energy.com>",]
license = "Apache-2.0"
readme = "README.md"
homepage="https://github.com/Reveal-Energy-Services/orchid-python-api"
repository="https://github.com/Reveal-Energy-Services/orchid-python-api"
documentation="https://github.com/Reveal-Energy-Services/orchid-python-api/wiki"
keywords=["Orchid",
          "Fracture Diagnostics"]
classifiers=[
    # Trove classifiers
    # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 4 - Beta",
    # Indicate who your project is intended for
    "Intended Audience :: Science/Research",
    "Intended Audience :: Other Audience",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
]
include=["LICENSE", "orchid/*.py", "orchid/VERSION", "orchid_python_api/examples/*.ipynb", "copy_orchid_examples.py"]
exclude=["orchid/.ipynb_checkpoints/", "**/*py~"]

[tool.poetry.dependencies]
python = "^3.8"
pythonnet = "^2.5.1"
numpy = "^1.19.0"
python-dateutil = "^2.8.1"
deal = "^3.9.0"
matplotlib = "^3.3.0"
seaborn = "^0.10.1"
pyyaml = "^5.3.1"
pandas = "^1.0.5"
typing-extensions = "^3.7.4.2"
toolz = "^0.10.0"
pint = "^0.16.1"

[tool.poetry.dev-dependencies]
sphinx = "^3.1.2"
ipython = "^7.16.1"
pyhamcrest = "^2.0.2"
behave = "^1.2.6"
jupyter = "^1.0.0"
jupyterlab = "^2.2.0"
scipy = "^1.5.1"
pipenv-setup = "^3.1.1"
wheel = "^0.34.2"
invoke = "^1.4.1"
toml = "^0.10.1"
grip = "^4.5.2"

[tool.poetry.scripts]
copy_orchid_examples = "copy_orchid_examples:main"

[tool.poetry.urls]
Issues = "https://github.com/Reveal-Energy-Services/orchid-python-api/issues"
